moved engine valid check to engine factory; adjusted unit tests
