+++ src/datagrunt/core/engines.py
@@ -377,6 +377,8 @@ class EngineFactory:
         self.db_table = DuckDBQueries(self.filepath).database_table_name
         if not os.path.exists(self.filepath):
             raise FileNotFoundError
+        if self.engine not in EngineProperties.valid_engines:
+            raise ValueError(EngineProperties.value_error_message.format(engine=self.engine))
 
     def create_reader(self):
         """Create a reader engine instance.
